Remove the rest of kas::theme to kas::draw and kas::traits
