+++ kas-wgpu/examples/counter.rs
@@ -9,8 +9,8 @@
 use kas::class::HasText;
 use kas::event::{Manager, VoidMsg, VoidResponse};
 use kas::macros::{make_widget, VoidMsg};
-use kas::theme::ThemeApi;
 use kas::widget::{Label, TextButton, Window};
+use kas::ThemeApi;
 
 #[derive(Clone, Debug, VoidMsg)]
 enum Message {
