+++ kas-wgpu/examples/custom-theme.rs
@@ -13,8 +13,8 @@ use kas::draw::Colour;
 use kas::event::{Manager, VoidMsg, VoidResponse};
 use kas::geom::Rect;
 use kas::macros::{make_widget, VoidMsg};
-use kas::theme::{ThemeAction, ThemeApi};
 use kas::widget::*;
+use kas::{ThemeAction, ThemeApi};
 use kas_theme::Theme;
 
 use kas_wgpu::draw::DrawPipe;
