+++ src/traits.rs
@@ -6,12 +6,13 @@
 //! Widget traits
 
 use std::fmt;
+use std::ops::DerefMut;
 use std::time::Duration;
 
+use crate::draw::{DrawHandle, SizeHandle};
 use crate::event::{Callback, CursorIcon, Handler, Manager, ManagerState, UpdateHandle, VoidMsg};
 use crate::geom::{Coord, Rect, Size};
 use crate::layout::{self, AxisInfo, SizeRules};
-use crate::theme::{DrawHandle, SizeHandle};
 use crate::{AlignHints, CoreData, WidgetId};
 
 pub trait CloneTo {
@@ -329,4 +330,53 @@ pub trait Window: Widget + Handler<Msg = VoidMsg> {
 
     /// Trigger a callback (see `iter_callbacks`).
     fn trigger_callback(&mut self, index: usize, mgr: &mut Manager);
+}
+
+/// Toolkit actions needed after theme adjustment, if any
+#[derive(Debug, Clone, Copy, Eq, PartialEq, Ord, PartialOrd)]
+pub enum ThemeAction {
+    /// No action needed
+    None,
+    /// All windows require redrawing
+    RedrawAll,
+    /// Theme sizes changed: must call [`Theme::update_window`] and resize
+    ThemeResize,
+}
+
+/// Interface through which a theme can be adjusted at run-time
+///
+/// All methods return a [`ThemeAction`] to enable correct action when a theme
+/// is updated via [`Manager::adjust_theme`]. When adjusting a theme before
+/// the UI is started, this return value can be safely ignored.
+///
+/// [`Manager::adjust_theme`]: crate::event::Manager::adjust_theme
+pub trait ThemeApi {
+    /// Set font size. Default is 18. Units are unknown.
+    fn set_font_size(&mut self, size: f32) -> ThemeAction;
+
+    /// Change the colour scheme
+    ///
+    /// If no theme by this name is found, the theme is unchanged.
+    // TODO: revise scheme identification and error handling?
+    fn set_colours(&mut self, _scheme: &str) -> ThemeAction;
+
+    /// Change the theme itself
+    ///
+    /// Themes may do nothing, or may react according to their own
+    /// interpretation of this method.
+    fn set_theme(&mut self, _theme: &str) -> ThemeAction {
+        ThemeAction::None
+    }
+}
+
+impl<T: ThemeApi> ThemeApi for Box<T> {
+    fn set_font_size(&mut self, size: f32) -> ThemeAction {
+        self.deref_mut().set_font_size(size)
+    }
+    fn set_colours(&mut self, scheme: &str) -> ThemeAction {
+        self.deref_mut().set_colours(scheme)
+    }
+    fn set_theme(&mut self, theme: &str) -> ThemeAction {
+        self.deref_mut().set_theme(theme)
+    }
 }
