+++ src/widget/radiobox.rs
@@ -10,11 +10,11 @@ use std::fmt::{self, Debug};
 
 use super::Label;
 use crate::class::HasBool;
+use crate::draw::{DrawHandle, SizeHandle};
 use crate::event::{Action, Handler, Manager, ManagerState, Response, UpdateHandle, VoidMsg};
 use crate::geom::Rect;
 use crate::layout::{AxisInfo, SizeRules};
 use crate::macros::Widget;
-use crate::theme::{DrawHandle, SizeHandle};
 use crate::{Align, AlignHints, CoreData, Layout, Widget, WidgetCore, WidgetId};
 
 /// A bare radiobox (no label)
